Nov 9 2025 10:02AM

+++ 11-Odd_or_Even.py
@@ -1,5 +1,5 @@
 number= input("Enter a number: ")
-r = int(number)%2
+r = float(number)%2
 
 if r == 0:
     print(f"{number} is even.")
